business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/GetParm.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class GetParm implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+p ");
        messageString.append(command.getParameter("Parm"));
        messageString.append("\n");
        return messageString;
    }
}
